services/notifications/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "notifications-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "flask==3.1.3",
  "flask-socketio==5.5.1",
  "gunicorn==25.3.0",
  "eventlet==0.40.4",
  "pyjwt==2.10.1",
  "redis[hiredis]==7.4.0",
  "protobuf==6.33.6",
  "pex==2.93.4",
  "grpcio-tools==1.80.0",
  "mypy-protobuf",
]

[project.optional-dependencies]
dev = ["mypy"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
explicit_package_bases = true
mypy_path = "src"
